v0.02 -  Creating Controls
This time, we started creating our app.php file. This file will control the operations with the database, validating accesses and operations. We also started to implement the css style files to adapt the user panel. Another thing that was added was a folder to store the operational modules that will contain the resources that we are going to use in our system.