v0.07 - Changes
This time we made some changes to our system to convert and eliminate the need to install the system to something simpler being necessary. After importing the database tables into the database we create on the server, we also added some small settings for us to use for connect the system with the database of our new server.